Revert "setup navigation"
This reverts commit 620a017bd909e20a1554ad713ef644c20e790ea2.
