--- app/src/main/kotlin/com/streafy/pizzashift2024/main/MainActivity.kt
@@ -1,70 +1,0 @@
-package com.streafy.pizzashift2024.main
-
-import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
-import androidx.activity.viewModels
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
-import com.github.terrakok.cicerone.NavigatorHolder
-import com.github.terrakok.cicerone.androidx.AppNavigator
-import com.streafy.pizzashift2024.R
-import com.streafy.pizzashift2024.databinding.ActivityMainBinding
-import dagger.hilt.android.AndroidEntryPoint
-import javax.inject.Inject
-
-@AndroidEntryPoint
-class MainActivity : AppCompatActivity() {
-
-    private lateinit var binding: ActivityMainBinding
-
-    private val viewModel: MainViewModel by viewModels()
-
-    @Inject lateinit var navigatorHolder: NavigatorHolder
-    private val navigator = AppNavigator(this, R.id.fragment_container)
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        binding = ActivityMainBinding.inflate(layoutInflater).also { setContentView(it.root) }
-        ViewCompat.setOnApplyWindowInsetsListener(binding.main) { v, insets ->
-            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
-            v.setPadding(systemBars.left, systemBars.top, systemBars.right, 0)
-            insets
-        }
-
-        setupBottomNavigation()
-    }
-
-    override fun onResumeFragments() {
-        super.onResumeFragments()
-        navigatorHolder.setNavigator(navigator)
-    }
-
-    override fun onPause() {
-        super.onPause()
-        navigatorHolder.removeNavigator()
-    }
-
-    private fun setupBottomNavigation() {
-        binding.bottomNavigation.setOnItemSelectedListener { item ->
-            when (item.itemId) {
-                R.id.pizza_menu_item -> {
-                    viewModel.openOption(NavigationOption.PIZZA_LIST)
-                    true
-                }
-                R.id.orders_menu_item -> {
-                    viewModel.openOption(NavigationOption.ORDERS)
-                    true
-                }
-                else -> false
-            }
-        }
-        binding.bottomNavigation.setOnItemReselectedListener { item ->
-            when (item.itemId) {
-                R.id.pizza_menu_item -> Unit
-                R.id.orders_menu_item -> Unit
-            }
-        }
-    }
-}
