--- app/src/main/kotlin/com/streafy/pizzashift2024/main/MainRouter.kt
@@ -1,12 +1,0 @@
-package com.streafy.pizzashift2024.main
-
-interface MainRouter {
-
-    fun openPizzaList()
-
-    fun openOrders()
-
-    fun openCart()
-
-    fun openProfile()
-}
