--- app/src/main/kotlin/com/streafy/pizzashift2024/orders/navigation/OrdersRouter.kt
@@ -1,6 +1,0 @@
-package com.streafy.pizzashift2024.orders.navigation
-
-interface OrdersRouter {
-
-    fun openDetails()
-}
