--- app/src/main/kotlin/com/streafy/pizzashift2024/orders/navigation/OrdersScreen.kt
@@ -1,8 +1,0 @@
-package com.streafy.pizzashift2024.orders.navigation
-
-import com.github.terrakok.cicerone.androidx.FragmentScreen
-import com.streafy.pizzashift2024.orders.presentation.OrdersFragment
-
-val ordersScreen = FragmentScreen {
-    OrdersFragment.newInstance()
-}
