--- app/src/main/kotlin/com/streafy/pizzashift2024/pizzalist/navigation/PizzaListRouter.kt
@@ -1,6 +1,0 @@
-package com.streafy.pizzashift2024.pizzalist.navigation
-
-interface PizzaListRouter {
-
-    fun openDetails()
-}
