--- app/src/main/kotlin/com/streafy/pizzashift2024/pizzalist/navigation/PizzaListScreen.kt
@@ -1,8 +1,0 @@
-package com.streafy.pizzashift2024.pizzalist.navigation
-
-import com.github.terrakok.cicerone.androidx.FragmentScreen
-import com.streafy.pizzashift2024.pizzalist.presentation.PizzaListFragment
-
-val pizzaListScreen = FragmentScreen {
-    PizzaListFragment.newInstance()
-}
